# tests/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentCatalogObjectStub": {
        "properties": {
          "capabilities": {
            "items": {
              "$ref": "#/components/schemas/Permission"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "host",
              "system",
              "network",
              "service",
              "credential_reference",
              "runbook",
              "decision",
              "project"
            ],
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "parent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentAssetNode"
              },
              {
                "$ref": "#/components/schemas/AgentAssetStubNode"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent"
          },
          "parent_path": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AgentAssetNode"
                },
                {
                  "$ref": "#/components/schemas/AgentAssetStubNode"
                }
              ]
            },
            "title": "Parent Path",
            "type": "array"
          },
          "placement_state": {
            "anyOf": [
              {
                "enum": [
                  "root",
                  "assigned",
                  "unassigned",
                  "unknown"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Placement State"
          },
          "ref": {
            "title": "Ref",
            "type": "string"
          },
          "visibility": {
            "const": "stub",
            "default": "stub",
            "title": "Visibility",
            "type": "string"
          }
        },
        "required": [
          "ref",
          "id",
          "kind",
          "label"
        ],
        "title": "AgentCatalogObjectStub",
        "type": "object"
      },
...
      "AgentContextOut": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "filters": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "boolean"
                }
              ]
            },
            "title": "Filters",
            "type": "object"
          },
          "kind": {
            "anyOf": [
              {
                "enum": [
                  "host",
                  "system",
                  "network",
                  "service",
                  "credential_reference",
                  "runbook",
                  "decision",
                  "project"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "objects": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AgentCatalogObjectContext"
                },
                {
                  "$ref": "#/components/schemas/AgentCatalogObjectStub"
                }
              ]
            },
            "title": "Objects",
            "type": "array"
          },
          "query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query"
          }
        },
        "required": [
          "count",
          "objects"
        ],
        "title": "AgentContextOut",
        "type": "object"
      },
...
      "AgentSearchOut": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "filters": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "boolean"
                }
              ]
            },
            "title": "Filters",
            "type": "object"
          },
          "kind": {
            "anyOf": [
              {
                "enum": [
                  "host",
                  "system",
                  "network",
                  "service",
                  "credential_reference",
                  "runbook",
                  "decision",
                  "project"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query"
          },
          "results": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AgentCatalogObjectSummary"
                },
                {
                  "$ref": "#/components/schemas/AgentCatalogObjectStub"
                }
              ]
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "count",
          "results"
        ],
        "title": "AgentSearchOut",
        "type": "object"
      },
...
      "ApiErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "required": [
          "error"
        ],
        "title": "ApiErrorResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Blockwart",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/agent/context": {
      "get": {
        "operationId": "agent_context_api_agent_context_get",
        "parameters": [
          {
            "description": "Search term for context retrieval",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search term for context retrieval",
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "host",
                    "system",
                    "network",
                    "service",
                    "credential_reference",
                    "runbook",
                    "decision",
                    "project"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "description": "Typed parent reference",
            "in": "query",
            "name": "parent",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Typed parent reference",
              "title": "Parent"
            }
          },
          {
            "description": "Resolved exact IP address",
            "in": "query",
            "name": "ip",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Resolved exact IP address",
              "title": "Ip"
            }
          },
          {
            "in": "query",
            "name": "port",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 65535,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Port"
            }
          },
          {
            "in": "query",
            "name": "endpoint_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Endpoint Type"
            }
          },
          {
            "in": "query",
            "name": "protocol",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Protocol"
            }
          },
          {
            "in": "query",
            "name": "exposure",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Exposure"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "lifecycle",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "planned",
                    "active",
                    "retired"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lifecycle"
            }
          },
          {
            "in": "query",
            "name": "health",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "unknown",
                    "healthy",
                    "degraded",
                    "down",
                    "maintenance"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Health"
            }
          },
          {
            "in": "query",
            "name": "source_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "unknown",
                    "manual",
                    "import",
                    "discovery"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Type"
            }
          },
          {
            "description": "Exact computed freshness state",
            "in": "query",
            "name": "stale",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact computed freshness state",
              "title": "Stale"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 5,
              "maximum": 20,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentContextOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Precondition failed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Precondition required"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Agent Context",
        "tags": [
          "agent-readonly"
        ]
      }
    },
    "/api/agent/objects/{object_id}": {
      "get": {
        "operationId": "agent_object_context_api_agent_objects__object_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "object_id",
            "required": true,
            "schema": {
              "title": "Object Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentContextOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Precondition failed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Precondition required"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Agent Object Context",
        "tags": [
          "agent-readonly"
        ]
      }
    },
    "/api/agent/search": {
      "get": {
        "operationId": "agent_search_api_agent_search_get",
        "parameters": [
          {
            "description": "Search term for id, label, summary, or data",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search term for id, label, summary, or data",
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "host",
                    "system",
                    "network",
                    "service",
                    "credential_reference",
                    "runbook",
                    "decision",
                    "project"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "description": "Typed parent reference",
            "in": "query",
            "name": "parent",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Typed parent reference",
              "title": "Parent"
            }
          },
          {
            "description": "Resolved exact IP address",
            "in": "query",
            "name": "ip",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Resolved exact IP address",
              "title": "Ip"
            }
          },
          {
            "in": "query",
            "name": "port",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 65535,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Port"
            }
          },
          {
            "in": "query",
            "name": "endpoint_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Endpoint Type"
            }
          },
          {
            "in": "query",
            "name": "protocol",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Protocol"
            }
          },
          {
            "in": "query",
            "name": "exposure",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Exposure"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "lifecycle",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "planned",
                    "active",
                    "retired"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lifecycle"
            }
          },
          {
            "in": "query",
            "name": "health",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "unknown",
                    "healthy",
                    "degraded",
                    "down",
                    "maintenance"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Health"
            }
          },
          {
            "in": "query",
            "name": "source_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "unknown",
                    "manual",
                    "import",
                    "discovery"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Type"
            }
          },
          {
            "description": "Exact computed freshness state",
            "in": "query",
            "name": "stale",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact computed freshness state",
              "title": "Stale"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentSearchOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Permission denied"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Precondition failed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Precondition required"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Agent Search",
        "tags": [
          "agent-readonly"
        ]
      }
    },
...
  }
}